Fix supervised and unsupervised forests
