+++ sktree/ensemble/_supervised_forest.py
@@ -161,9 +161,15 @@ class ObliqueRandomForestClassifier(ForestClassifier):
         - If float, then draw `max_samples * X.shape[0]` samples. Thus,
           `max_samples` should be in the interval `(0.0, 1.0]`.
 
-    feature_combinations : float, default=1.5
+    feature_combinations : float, default=None
         The number of features to combine on average at each split
-        of the decision trees.
+        of the decision trees. If ``None``, then will default to the minimum of
+        ``(1.5, n_features)``. This controls the number of non-zeros is the
+        projection matrix. Setting the value to 1.0 is equivalent to a
+        traditional decision-tree. ``feature_combinations * max_features``
+        gives the number of expected non-zeros in the projection matrix of shape
+        ``(max_features, n_features)``. Thus this value must always be less than
+        ``n_features`` in order to be valid.
 
     Attributes
     ----------
@@ -290,7 +296,7 @@ class ObliqueRandomForestClassifier(ForestClassifier):
         warm_start=False,
         class_weight=None,
         max_samples=None,
-        feature_combinations=1.5,
+        feature_combinations=None,
     ):
         super().__init__(
             estimator=ObliqueDecisionTreeClassifier(),
@@ -610,7 +616,7 @@ class PatchObliqueRandomForestClassifier(ForestClassifier):
         min_patch_height=1,
         max_patch_height=1,
         min_patch_width=1,
-        max_patch_width=5,
+        max_patch_width=1,
         data_height=1,
         data_width=None,
     ):
@@ -627,7 +633,6 @@ class PatchObliqueRandomForestClassifier(ForestClassifier):
                 "max_leaf_nodes",
                 "min_impurity_decrease",
                 "random_state",
-                "feature_combinations",
                 "min_patch_height",
                 "max_patch_height",
                 "min_patch_width",
