+++ sktree/tests/meson.build
@@ -1,6 +1,7 @@
 python_sources = [
   '__init__.py',
-  'test_forest.py',
+  'test_supervised_forest.py',
+  'test_unsupervised_forest.py',
 ]
 
 py3.install_sources(
