+++ sktree/tree/_unsup_tree.pyx
@@ -84,7 +84,7 @@ cdef class UnsupervisedTreeBuilder:
         self,
         UnsupervisedTree tree,
         object X,
-        cnp.ndarray sample_weight=None
+        const DOUBLE_t[:] sample_weight=None
     ):
         """Build a decision tree from the training set X."""
         pass
@@ -190,7 +190,7 @@ cdef class UnsupervisedBestFirstTreeBuilder(UnsupervisedTreeBuilder):
         self,
         UnsupervisedTree tree,
         object X,
-        cnp.ndarray sample_weight=None
+        const DOUBLE_t[:] sample_weight=None
     ):
         """Build a decision tree from the training set X."""
         # check input
@@ -401,7 +401,7 @@ cdef class UnsupervisedDepthFirstTreeBuilder(UnsupervisedTreeBuilder):
         self,
         UnsupervisedTree tree,
         object X,
-        cnp.ndarray sample_weight=None
+        const DOUBLE_t[:] sample_weight=None
     ):
         """Build a decision tree from the training set (X, y)."""
 
@@ -713,11 +713,15 @@ cdef class UnsupervisedTree(BaseTree):
             expected_shape=value_shape
         )
 
+        self.capacity = node_ndarray.shape[0]
+        if self._resize_c(self.capacity) != 0:
+            raise MemoryError("resizing tree to %d" % self.capacity)
+
         cdef Node[::1] node_memory_view = node_ndarray
-        cdef DOUBLE_t[:, :, ::1] value_memory_view = value_ndarray
+        cdef DOUBLE_t[:] value_memory_view = value_ndarray
         nodes = memcpy(self.nodes, &node_memory_view[0],
                        self.capacity * sizeof(Node))
-        value = memcpy(self.value, &value_memory_view[0, 0, 0],
+        value = memcpy(self.value, &value_memory_view[0],
                        self.capacity * self.value_stride * sizeof(double))
 
     cdef int _set_split_node(
